requirements.txt
boto3==1.26.137
botocore==1.29.137
psycopg2-binary==2.9.6
requests==2.31.0
python-dateutil==2.8.2
six==1.16.0
urllib3==1.26.18
